requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anchorpy==0.20.1
anchorpy-core==0.2.0
anyio==4.4.0
apischema==0.17.5
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
base58==2.1.1
based58==0.1.1
beautifulsoup4==4.12.3
bitarray==2.9.2
bleach==6.1.0
borsh-construct==0.1.0
cachetools==4.2.4
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
ckzg==1.0.2
click==8.1.7
colorlog==6.8.2
comm==0.2.2
construct==2.10.68
construct-typing==0.5.6
contourpy==1.2.1
cycler==0.12.1
cytoolz==0.12.3
dateparser==1.2.0
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.3.1
eth-utils==4.1.1
eth_abi==5.1.0
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.20.0
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
grpcio==1.64.1
h11==0.14.0
hexbytes==0.3.1
httpcore==0.16.3
httpx==0.23.3
idna==3.7
ipykernel==6.29.4
ipython==8.26.0
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jito_searcher_client==0.1.4
json5==0.9.25
jsonalias==0.1.1
jsonpointer==3.0.0
jsonrpcclient==4.0.3
jsonrpcserver==5.0.9
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
kiwisolver==1.4.5
lru-dict==1.2.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mistune==3.0.2
more-itertools==8.14.0
multidict==6.0.5
Naked==0.1.32
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.1
notebook_shim==0.2.4
numpy==2.0.0
OSlash==0.6.3
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parsimonious==0.10.0
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pycryptodome==3.20.0
Pygments==2.18.0
pyheck==0.1.5
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
pyunormalize==15.1.0
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rlp==4.0.1
rpds-py==0.18.1
Send2Trash==1.8.3
shellescape==3.8.1
six==1.16.0
sniffio==1.3.1
solana==0.34.2
solders==0.21.0
soupsieve==2.5
stack-data==0.6.3
sumtypes==0.1a6
terminado==0.18.1
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
tornado==6.4.1
traitlets==5.14.3
types-cachetools==4.2.10
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uri-template==1.3.0
urllib3==2.2.2
wcwidth==0.2.13
web3==6.20.0
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
wrapt==1.16.0
yarl==1.9.4
zetamarkets_py==0.2.61
